requirements-dev.txt
factory_boy==3.2.0
pytest-django==4.4.0
